Complete request CreateNewPet and DeletesPet
