+++ petStoreKarate/petStoreKarate/.gitignore
@@ -1,0 +1,4 @@
+#ignore .idea
+.idea
+#ignore target
+/target/
