+++ petStoreKarate/petStoreKarate/src/test/java/karate/request/DeletesPet.java
@@ -1,0 +1,10 @@
+package karate.request;
+
+import com.intuit.karate.junit5.Karate;
+
+public class DeletesPet {
+    @Karate.Test
+    Karate testSample(){
+        return Karate.run("classpath:karate/request/DeletesPet.feature").relativeTo(getClass());
+    }
+}
