+++ petStoreKarate/petStoreKarate/src/test/resources/karate-properties.json
@@ -1,0 +1,6 @@
+{
+  "dev": {
+    "url": "https://petstore.swagger.io/v2/",
+    "petId": "5"
+  }
+}
