+++ petStoreKarate/petStoreKarate/src/test/resources/karate/request/responseCreateOk.json
@@ -1,0 +1,18 @@
+{
+  "id": "#number",
+  "category": {
+  "id": "#number",
+  "name": "#string"
+},
+  "name": "#string",
+  "photoUrls": [
+    "#string"
+  ],
+  "tags": [
+  {
+  "id": "#number",
+  "name": "#string"
+}
+],
+ "status": "#string"
+}
