+++ petStoreKarate/petStoreKarate/src/test/resources/karate/request/responseErrorFilterPetById.json
@@ -1,0 +1,5 @@
+{
+  "code": "#number",
+  "type": "#string",
+  "message": "#string"
+}
