+++ petStoreKarate/petStoreWithKarate/src/test/java/karate/request/DeletesPets.java
@@ -1,0 +1,10 @@
+package karate.request;
+
+import com.intuit.karate.junit5.Karate;
+
+public class DeletesPetsPositive {
+    @Karate.Test
+    Karate testSample(){
+        return Karate.run("classpath:karate.request/DeletesPetsPositive.feature").relativeTo(getClass());
+    }
+}
