--- petStoreKarate/petStoreWithKarate/src/test/java/karate/request/DeletesPetsPositive.java
@@ -1,4 +1,0 @@
-package karate.request;
-
-public class DeletesPetsPositive {
-}
